Added USER_TYPE.SUBSCRIBER (as per request) Added User.getUserName (retrieves the twitch login name) Changed USER_TYPE.value for most USER_TYPES (to make space for SUBSCRIBER in between DEFAULT and STAFF) Changed User.userID to long Modified User.getName -> User.getDisplayName (see above)
Depricated TwirkListener.onWhisper (twitch uses another wisper system now)
Depricated TwirkListener.onMode (check users for User.isMod instead, it is unreliable anyway)

+++ twirc/src/main/java/com/gikk/twirk/types/usernotice/Usernotice.java
@@ -1,11 +1,10 @@
 package com.gikk.twirk.types.usernotice;
 
-import java.util.List;
-
 import com.gikk.twirk.enums.USER_TYPE;
 import com.gikk.twirk.types.AbstractType;
 import com.gikk.twirk.types.emote.Emote;
 import com.gikk.twirk.types.emote.EmoteImpl;
+import java.util.List;
 
 public interface Usernotice extends AbstractType {
 	
@@ -64,7 +63,7 @@ public interface Usernotice extends AbstractType {
 	 * 
 	 * @return The users unique user ID
 	 */
-	public int getUserID();
+	public long getUserID();
 	
 	/**Retrieves the re-subscribing users login name. At the moment, this is the users display name in lower cases. It might
 	 * change in the future, if Twitch implements name changes.
